+++ publiccms-by-gradle/src/com/publiccms/views/controller/web/changyan/ChangyanController.java
@@ -87,9 +87,12 @@ public class ChangyanController extends AbstractController {
     public MappingJacksonValue login(String callback, HttpServletRequest request, HttpServletResponse response) {
         Cookie userCookie = getCookie(request.getCookies(), COOKIES_USER);
         if (null != userCookie && notEmpty(userCookie.getValue())) {
-            String[] userData = userCookie.getValue().split(COOKIES_USER_SPLIT);
-            if (userData.length > 1) {
-                sysUserTokenService.delete(userData[1]);
+            String value = userCookie.getValue();
+            if (null != value) {
+                String[] userData = value.split(COOKIES_USER_SPLIT);
+                if (userData.length > 1) {
+                    sysUserTokenService.delete(userData[1]);
+                }
             }
         }
         clearUserToSession(request.getContextPath(), request.getSession(), response);
